+++ src/main/LexerDemo.java
@@ -9,40 +9,11 @@ import java.io.IOException;
 import java.util.ArrayList;
 public class LexerDemo {
 	public static void main(String[] args) throws IOException {
-//		final long startTime = System.nanoTime();
-//		ArrayList<Token> tokens = new Lexer().lex("in/String.java");
-//		final long stopTime = System.nanoTime();
-//		demo2(tokens);
-//		System.out.println((stopTime - startTime) / 1e6 + "ms");
-
-		ArrayList<Token> tokens ;
-	}
-
-	public static class OtherDemo extends Application {
-		public static void main(String[] args) {
-			launch(args);
-		}
-		@Override
-		public void start(Stage primaryStage) {
-			ArrayList<Token> tokens;
-			try {
-				tokens = new Lexer().lex("in/MyClass.java");
-			} catch (IOException e) {
-				return;
-			}
-			ParserV1 parser = new ParserV1(tokens);
-			TreeNode tree = parser.parse();
-			System.out.println(tree);
-
-			TreeView<String> treeView = new TreeView<>(parser.debugRoot);
-			primaryStage.setScene(new Scene(new Group(treeView), 900, 900));
-
-
-			treeView.prefWidthProperty().bind(primaryStage.widthProperty());
-			treeView.prefHeightProperty().bind(primaryStage.heightProperty());
-
-			primaryStage.show();
-		}
+		final long startTime = System.nanoTime();
+		ArrayList<Token> tokens = new Lexer().lex("in/String.java");
+		final long stopTime = System.nanoTime();
+		demo2(tokens);
+		System.out.println((stopTime - startTime) / 1e6 + "ms");
 	}
 
 	private static void demo1(ArrayList<Token> tokens) {
